feat: :rocket: 勝利判定をコンソールからシーンへ移行
