+++ Assets/ML-Agents/Timers/Main_Standard_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":27.0975328,"total":27.2645721,"children":{"InitializeActuators":{"count":1,"self":0.0030007,"total":0.0030007,"children":null},"InitializeSensors":{"count":1,"self":0.0010003,"total":0.0010003,"children":null},"AgentSendState":{"count":1132,"self":0.0074437999999999995,"total":0.0308505,"children":{"CollectObservations":{"count":1,"self":0.0020894,"total":0.0020894,"children":null},"WriteActionMask":{"count":1,"self":0,"total":0,"children":null},"RequestDecision":{"count":1,"self":0.021317299999999997,"total":0.021317299999999997,"children":null}}},"DecideAction":{"count":1132,"self":0.1286558,"total":0.1286558,"children":null},"AgentAct":{"count":1132,"self":0.0010054999999999999,"total":0.0010054999999999999,"children":null}},"gauges":{},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1727194456","unity_version":"2022.3.23f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2022.3.23f1\\Editor\\Unity.exe -projectpath C:\\Users\\baske\\UnityProjects\\3DConnectFour -useHub -hubIPC -cloudEnvironment production","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"3.0.0-exp.1","scene_name":"Main_Standard","end_time_seconds":"1727194483"}}
+{"count":1,"self":33.733964799999995,"total":33.793316499999996,"children":{"InitializeActuators":{"count":1,"self":0.0010007,"total":0.0010007,"children":null},"InitializeSensors":{"count":1,"self":0.0005177,"total":0.0005177,"children":null},"AgentSendState":{"count":1459,"self":0.0055203,"total":0.0111296,"children":{"CollectObservations":{"count":2,"self":0,"total":0,"children":null},"WriteActionMask":{"count":2,"self":0,"total":0,"children":null},"RequestDecision":{"count":2,"self":0.0056092999999999994,"total":0.0056092999999999994,"children":null}}},"DecideAction":{"count":1459,"self":0.0457029,"total":0.0457029,"children":null},"AgentAct":{"count":1459,"self":0,"total":0,"children":null}},"gauges":{},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1727237196","unity_version":"2022.3.23f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2022.3.23f1\\Editor\\Unity.exe -projectpath C:\\Users\\baske\\UnityProjects\\3DConnectFour -useHub -hubIPC -cloudEnvironment production","communication_protocol_version":"1.5.0","com.unity.ml-agents_version":"3.0.0-exp.1","scene_name":"Main_Standard","end_time_seconds":"1727237230"}}
